nyc-traffic-cam/server/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
httpx==0.27.2
aiosqlite==0.20.0
pillow==11.0.0
numpy==2.1.3
